// dub-packages/prisma-oppulence/generated/zod/outputTypeSchemas/LeadFindManyArgsSchema.ts
export const LeadFindManyArgsSchema: z.ZodType<Prisma.LeadFindManyArgs> = z.object({
  select: LeadSelectSchema.optional(),
  include: LeadIncludeSchema.optional(),
  where: LeadWhereInputSchema.optional(),
  orderBy: z.union([ LeadOrderByWithRelationInputSchema.array(),LeadOrderByWithRelationInputSchema ]).optional(),
  cursor: LeadWhereUniqueInputSchema.optional(),
  take: z.number().optional(),
  skip: z.number().optional(),
  distinct: z.union([ LeadScalarFieldEnumSchema,LeadScalarFieldEnumSchema.array() ]).optional(),
}).strict() ;